Separate extensions into their own files
